fix(chat): resolve subscription cancellation for users with multiple plans
- Filter out free tier subscriptions to only cancel paid plans
- Refactor billing management into separate components for better maintainability
- Fix subscription cancellation failing due to attempting to cancel free tier
- Add proper filtering logic to distinguish between free and paid subscriptions
- Update isCanceled detection to work with filtered paid subscription items
- Resolve TypeScript and ESLint issues with Clerk experimental hooks
